moved CvUnitSelectionCriteria class to separate file (#1243)
