+++ Sources/C2C (VS2019).vcxproj
@@ -284,6 +284,7 @@
     <ClCompile Include="CvUnitGrouping.cpp" />
     <ClCompile Include="CvUnitInfo.cpp" />
     <ClCompile Include="CvUnitList.cpp" />
+    <ClCompile Include="CvUnitSelectionCriteria.cpp" />
     <ClCompile Include="CvUnitSort.cpp" />
     <ClCompile Include="CvViewport.cpp" />
     <ClCompile Include="CvXMLLoadUtility.cpp" />
@@ -472,6 +473,7 @@
     <ClInclude Include="CvUnitFilters.h" />
     <ClInclude Include="CvUnitGrouping.h" />
     <ClInclude Include="CvUnitList.h" />
+    <ClInclude Include="CvUnitSelectionCriteria.h" />
     <ClInclude Include="CvUnitSort.h" />
     <ClInclude Include="CvXMLLoadUtility.h" />
     <ClInclude Include="CvXMLLoadUtilityModTools.h" />
