+++ Sources/CvCityAI.cpp
@@ -18,6 +18,7 @@
 #include "CvSelectionGroup.h"
 #include "CvTeamAI.h"
 #include "CvUnit.h"
+#include "CvUnitSelectionCriteria.h"
 #include "CvDLLFAStarIFaceBase.h"
 #include "OutputRatios.h"
 #include "PlotInfo.h"
