+++ Sources/CvContractBroker.cpp
@@ -2,6 +2,7 @@
 
 #include "CvGameCoreDLL.h"
 #include "CvCity.h"
+#include "CvContractBroker.h"
 #include "CvGameAI.h"
 #include "CvGlobals.h"
 #include "CvMap.h"
