+++ Sources/CvDLLButtonPopup.cpp
@@ -21,6 +21,7 @@
 #include "CvSelectionGroup.h"
 #include "CvTeamAI.h"
 #include "CvUnit.h"
+#include "CvUnitSelectionCriteria.h"
 
 // Public Functions...
